+++ cfdb/__init__.py
@@ -13,4 +13,4 @@ try:
 except (ImportError, AttributeError):
     pass
 
-__version__ = '0.9.0'
+__version__ = '0.9.1'